Refactor: Move cache_response decorator to utils/cache.py
