+++ routes/main_routes.py
@@ -10,7 +10,7 @@ from flask_login import login_required, current_user
 from models import Player, Alliance, Event, MVPAssignment, WinnerAssignment, Blacklist, Guide
 from database import db
 from database_manager import query_user_data, get_user_data_by_id, get_user_data_optimized
-from app import cache_response
+from utils.cache import cache_response
 
 # Create blueprint for main routes
 bp = Blueprint('main', __name__)
